feat(ingest):move incremental lineage helper,auto-detect whether CLL exists
