+++ metadata-ingestion/tests/unit/api/source_helpers/test_incremental_lineage_helper.py
@@ -6,7 +6,7 @@ import pytest
 import datahub.metadata.schema_classes as models
 from datahub.emitter.mce_builder import make_dataset_urn, make_schema_field_urn
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
-from datahub.ingestion.api.source_helpers import auto_incremental_lineage
+from datahub.ingestion.api.incremental_lineage_helper import auto_incremental_lineage
 from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.ingestion.sink.file import write_metadata_file
 from tests.test_helpers import mce_helpers
@@ -88,7 +88,6 @@ def test_incremental_table_lineage(tmp_path, pytestconfig):
     processed_wus = auto_incremental_lineage(
         graph=None,
         incremental_lineage=True,
-        include_column_level_lineage=False,
         stream=[
             MetadataChangeProposalWrapper(
                 entityUrn=urn, aspect=aspect, systemMetadata=system_metadata
@@ -146,7 +145,6 @@ def test_incremental_column_level_lineage(
     processed_wus = auto_incremental_lineage(
         graph=mock_graph,
         incremental_lineage=True,
-        include_column_level_lineage=True,
         stream=[
             MetadataChangeProposalWrapper(
                 entityUrn=dataset_urn,
@@ -184,7 +182,6 @@ def test_incremental_column_lineage_less_upstreams_in_gms_aspect(
     processed_wus = auto_incremental_lineage(
         graph=mock_graph,
         incremental_lineage=True,
-        include_column_level_lineage=True,
         stream=[
             MetadataChangeProposalWrapper(
                 entityUrn=urn, aspect=aspect, systemMetadata=system_metadata
@@ -227,7 +224,6 @@ def test_incremental_column_lineage_more_upstreams_in_gms_aspect(
     processed_wus = auto_incremental_lineage(
         graph=mock_graph,
         incremental_lineage=True,
-        include_column_level_lineage=True,
         stream=[
             MetadataChangeProposalWrapper(
                 entityUrn=urn, aspect=aspect, systemMetadata=system_metadata
